requirements.txt
asgiref==3.5.0
colorama==0.4.4
Django~=3.2.11
django-cors-headers==3.11.0
django-filter==21.1
django-tree-queries==0.8.0
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
djangorestframework-simplejwt==5.0.0
# Pillow==9.0.0
PyJWT==2.3.0
pyotp==2.6.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
gunicorn==20.1.0
requests~=2.28.1
django4-background-tasks==1.2.7
xhtml2pdf~=0.2.9
XlsxWriter~=3.1.0
xlwt~=1.3.0
paho-mqtt==2.1.0